Move custom matcher in helper file.
