+++ spec/ultimateGuitarScraperSpec.js
@@ -1,33 +1,8 @@
 /* eslint-env jasmine */
 const utils = require('../lib/utils')
 const ugs = require('../lib/index')
-const Ajv = require('ajv')
-const ajv = new Ajv()
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 20000
-
-// TODO: Move this in helper file?
-const jsonSchemaMatcher = {
-  toMatchJsonSchema: (util, customEqualityTesters) => {
-    return {
-      compare: (actual, expected) => {
-        const schema = require(`./support/schemas/${expected}.json`)
-        const valid = ajv.validate(schema, actual)
-
-        if (!valid) console.error(ajv.errors)
-
-        const result = {}
-        result.pass = valid === true
-        if (result.pass) {
-          result.message = 'Expected ' + JSON.stringify(actual, null, 2) + `not to match schema ${expected}.json`
-        } else {
-          result.message = 'Expected ' + JSON.stringify(actual, null, 2) + `to match schema ${expected}.json`
-        }
-        return result
-      }
-    }
-  }
-}
 
 function basicSearchQuery () {
   return {
@@ -124,10 +99,6 @@ describe('utils', () => {
 })
 
 describe('ultimate-guitar-scraper', () => {
-  beforeEach(() => {
-    jasmine.addMatchers(jsonSchemaMatcher)
-  })
-
   describe('search', () => {
     it('searches TABs', (done) => {
       let query = basicSearchQuery()
